pyproject.toml
[project]
name = "recce"
version = "0.1.0"
description = "End-to-End Reconstruction-Classification Learning for Face Forgery Detection"
authors = [
    { name = "Anne Sulit", email = "92159716+axsulit@users.noreply.github.com" }
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "albumentations==1.0.3",
    "timm>=0.5.4",
    "tensorboardX==2.1",
    "scipy>=1.7.0",
    "PyYAML==5.3.1",
    "tqdm>=4.65.0",
    "protobuf==3.20.0",
    "scikit-learn>=1.0.0",
]
readme = "README.md"
requires-python = ">=3.8,<3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/recce"]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
